chore: move schema to separate file
